chore(time-sync): split context core + hook for fast refresh; add psycopg v3 dependency
