+++ frontend/src/contexts/useServerTime.ts
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+import { Ctx } from './ServerTimeContext';
+import type { ServerTimeContextValue } from './ServerTimeContext';
+
+export function useServerTime() {
+    return useContext(Ctx) as ServerTimeContextValue | null;
+}
